LLM-TRAINING/AI_AGENTS_POCS/FLIGHT-ASSISTANT/flight-bot/flight-backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv

# LangChain stack
langchain
langgraph
langchain-openai
langchain-community

# Vector DB
chromadb

# APIs
tavily-python
amadeus

